feat(subject): multiple controllers
Giving the option to have multiple controllers for a subject. This will
aid developers that want to use a subject in multiple locations. Only
when *all* controllers have closed will the close signal be passed to
the base controller.

BREAKING CHANGE: You must control a subject with the `control()` method.