fix: correct exporting for commonjs
CommonJS has been broken until now. Files with an extension of .js had
to be imported, inside an ES module. All files in the dist directory
were .js files. We've now moved the common/*.js to common/*.cjs
to get around this. In turn we've had to make sure that no extension is
specified in typescript and rely heavily upon node.js's conditional
importing.